fix all bf16

+++ src/maxdiffusion/models/embeddings_flax.py
@@ -70,9 +70,9 @@ class FlaxTimestepEmbedding(nn.Module):
 
     @nn.compact
     def __call__(self, temb):
-        temb = nn.Dense(self.time_embed_dim, dtype=self.dtype, name="linear_1")(temb)
+        temb = nn.Dense(self.time_embed_dim, dtype=self.dtype, param_dtype=self.dtype, name="linear_1")(temb)
         temb = nn.silu(temb)
-        temb = nn.Dense(self.time_embed_dim, dtype=self.dtype, name="linear_2")(temb)
+        temb = nn.Dense(self.time_embed_dim, dtype=self.dtype, param_dtype=self.dtype, name="linear_2")(temb)
         return temb
